DTOs y mas servicio

+++ src/main/java/com/gesinv/gestioninventario/repository/IProveedorRepo.java
@@ -3,7 +3,7 @@ package com.gesinv.gestioninventario.repository;
 import com.gesinv.gestioninventario.modelo.Proveedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IProveedorRepo extends JpaRepository<Proveedor, Long> {
+public interface IProveedorRepo extends JpaRepository<Proveedor, Integer> {
 
 }
